Re-organized folders and made demo page look nicer for github pages.
