+++ demo/src/routes/+layout.svelte
@@ -1,0 +1,18 @@
+<script>
+	let { children } = $props();
+  </script>
+  
+  
+  <div class="container">
+
+	<header>
+		<a target="_blank" href="https://github.com/lzinga/mdsvex-enhanced-images">mdsvex-enhanced-images</a>
+	  </header>
+	<div class="content">
+  
+	  <main>
+		{@render children()}
+	  </main>
+  
+	</div>
+  </div>
