+++ demo/src/routes/img with space.png
