+++ demo/src/routes/vitejs-logo.png
